backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
pytest>=7.4.3
httpx>=0.23.0,<0.24.0
solana==0.30.0
web3==6.11.1
requests>=2.31.0 
httpx>=0.25.0
boto3>=1.34.0 
